pyproject.toml
[project]
name = "yyshelp"
version = "0.1.0"
description = "yyshelp"
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-send-anything-anywhere>=0.6.1",
    "nonebot2>=2.2.1",
    "pillow>=10.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_saa"]
plugin_dirs = ["plugin"]
builtin_plugins = []
